feat: simplified crt usage (#9)
* simplified

* apply_lt_isolate

* feat: update test_isolate_coeffs_circuit

* feat: update

* limbs

* simplified input generation for crt format

* `lt_isolate`

* BigUint for value

* Add input functions to gadgets in crt (#10)

* Add input_u64 to BigUintPoly

* Add input_u64 to MontgomeryPoly

* Adding input_biguint to CRTPoly

* Fix tests in crt/mod.rs

* Change the LUT definition

* FIx isolate gadget

* Redesign input functions in gadget and add PackedCrtContext

* Add tests for PackedCrtPoly

---------


---------

+++ src/circuit/serde.rs
@@ -18,7 +18,7 @@ pub enum SerializablePolyGateType {
     Add,
     Sub,
     Mul,
-    Rotate { shift: usize },
+    Rotate { shift: i32 },
     PubLut { lookup_id: usize },
 }
